third_party/blink/web_tests/platform/generic/fast/dom/TreeWalker/previousSiblingLastChildSkip-expected.txt

Test for a specific problem with previousSibling that failed in older versions of WebKit.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS walker.firstChild(); walker.currentNode.id is 'A1'
PASS walker.nextNode(); walker.currentNode.id is 'C1'
PASS walker.nextNode(); walker.currentNode.id is 'C2'
PASS walker.nextNode(); walker.currentNode.id is 'D1'
PASS walker.nextNode(); walker.currentNode.id is 'D2'
PASS walker.nextNode(); walker.currentNode.id is 'B2'
PASS walker.previousSibling(); walker.currentNode.id is 'C2'
PASS successfullyParsed is true

TEST COMPLETE
